requirement.txt
python-dotenv
pandas
pandas-ta
ccxt
